your-thesis-project/requirements.txt
Flask
psycopg2-binary
pandas
uuid
pymongo
Flask-Cors
redis